+++ apps/desktop/src/i18n/types.ts
@@ -239,6 +239,7 @@ export interface Translations {
   keybinds: {
     title: string
     subtitle: (open: string) => string
+    search: string
     rebind: string
     reset: string
     resetAll: string
@@ -273,6 +274,7 @@ export interface Translations {
       providerApiKeys: string
       gateway: string
       apiKeys: string
+      keybinds: string
       keysTools: string
       keysSettings: string
       mcp: string
