+++ apps/desktop/src/lib/icons.ts
@@ -48,6 +48,7 @@ import {
   IconHelpCircle as HelpCircle,
   IconPhoto as ImageIcon,
   IconInfoCircle as Info,
+  IconKeyboard as Keyboard,
   IconKey as KeyRound,
   IconLayersIntersect2 as Layers3,
   IconLayoutDashboard as LayoutDashboard,
@@ -165,6 +166,7 @@ export {
   HelpCircle,
   ImageIcon,
   Info,
+  Keyboard,
   KeyRound,
   Layers3,
   LayoutDashboard,
